parser bot icloud/competitor_price_monitor/requirements.txt
beautifulsoup4>=4.12,<5
PyYAML>=6.0,<7
requests>=2.32,<3
playwright>=1.53,<2
gspread>=6.1,<7
google-auth>=2.39,<3
python-telegram-bot>=21.0,<22
